Created the Collaborators Section
The section that displays all the logos of the companies that Enco has collaborated with has been created.

A simple Materalize grid system has been used to display it . A container has also been used to concentrate the content in the middle.

Variables have also been added in the _variables.scss file to change them quickly if needed.