Removed Old Files.
The page has been restructured and the old files aren't needed anymore.

The main index file now becomes the index2.html (but renamed) the main CSS file has been removed and from now on we are going to use SASS to edit the CSS of the page, for a better understanding of it.